+++ admin/partials/mailchimp-woocommerce-admin-tabs.php
@@ -24,8 +24,13 @@ if (isset($options['mailchimp_api_key']) && $handler->hasValidApiKey()) {
         $show_sync_tab = true;
     }
 }
-
 ?>
+
+<?php if (!defined('PHP_VERSION_ID') || (PHP_VERSION_ID < 50600)): ?>
+    <div class="error notice is-dismissable">
+        <p><?php _e('MailChimp says: Please upgrade your PHP version to a minimum of 5.6', 'mailchimp-woocommerce'); ?></p>
+    </div>
+<?php endif; ?>
 
 <!-- Create a header in the default WordPress 'wrap' container -->
 <div class="wrap">