+++ includes/api/class-mailchimp-woocommerce-transform-products.php
@@ -40,7 +40,7 @@ class MailChimp_WooCommerce_Transform_Products
 
     /**
      * @param WP_Post $post
-     * @return MailChimp_Product
+     * @return MailChimp_WooCommerce_Product
      */
     public function transform(WP_Post $post)
     {
@@ -52,7 +52,7 @@ class MailChimp_WooCommerce_Transform_Products
 
         $is_variant = count($variants) > 1;
 
-        $product = new MailChimp_Product();
+        $product = new MailChimp_WooCommerce_Product();
 
         $product->setId($woo->get_id());
         $product->setHandle($post->post_name);
@@ -64,7 +64,7 @@ class MailChimp_WooCommerce_Transform_Products
 
         foreach ($variants as $variant) {
 
-            $product_variant = $this->variant($is_variant, $variant);
+            $product_variant = $this->variant($is_variant, $variant, $woo->get_title());
 
             $product_variant_title = $product_variant->getTitle();
 
@@ -87,9 +87,10 @@ class MailChimp_WooCommerce_Transform_Products
     /**
      * @param $is_variant
      * @param WP_Post $post
-     * @return MailChimp_ProductVariation
+     * @param string $fallback_title
+     * @return MailChimp_WooCommerce_ProductVariation
      */
-    public function variant($is_variant, $post)
+    public function variant($is_variant, $post, $fallback_title = null)
     {
         if ($post instanceof WC_Product || $post instanceof WC_Product_Variation) {
             $woo = $post;
@@ -101,21 +102,35 @@ class MailChimp_WooCommerce_Transform_Products
             }
         }
 
-        $variant = new MailChimp_ProductVariation();
+        $variant = new MailChimp_WooCommerce_ProductVariation();
 
         $variant->setId($woo->get_id());
         $variant->setUrl($woo->get_permalink());
-        $variant->setTitle($woo->get_title());
         $variant->setBackorders($woo->backorders_allowed());
         $variant->setImageUrl(get_the_post_thumbnail_url($post));
-        $variant->setInventoryQuantity(($woo->managing_stock() ? $woo->get_stock_quantity() : 0));
+        $variant->setInventoryQuantity(($woo->managing_stock() ? $woo->get_stock_quantity() : 1));
         $variant->setPrice($woo->get_price());
         $variant->setSku($woo->get_sku());
 
         if ($woo instanceof WC_Product_Variation) {
+
+            $variation_title = $woo->get_title();
+            if (empty($variation_title)) $variation_title = $fallback_title;
+
+            $title = array($variation_title);
+
+            foreach ($woo->get_variation_attributes() as $attribute => $value) {
+                if (is_string($value)) {
+                    $name = ucfirst(str_replace(array('attribute_pa_', 'attribute_'), '', $attribute));
+                    $title[] = "$name = $value";
+                }
+            }
+
+            $variant->setTitle(implode(' :: ', $title));
             $variant->setVisibility(($woo->variation_is_visible() ? 'visible' : ''));
         } else {
             $variant->setVisibility(($woo->is_visible() ? 'visible' : ''));
+            $variant->setTitle($woo->get_title());
         }
 
         return $variant;
@@ -166,7 +181,7 @@ class MailChimp_WooCommerce_Transform_Products
 
     /**
      * @param $id
-     * @return MailChimp_Product
+     * @return MailChimp_WooCommerce_Product
      */
     public static function deleted($id)
     {
@@ -174,12 +189,12 @@ class MailChimp_WooCommerce_Transform_Products
         $api = mailchimp_get_api();
 
         if (!($product = $api->getStoreProduct($store_id, "deleted_{$id}"))) {
-            $product = new MailChimp_Product();
+            $product = new MailChimp_WooCommerce_Product();
 
             $product->setId("deleted_{$id}");
             $product->setTitle("deleted_{$id}");
 
-            $variant = new MailChimp_ProductVariation();
+            $variant = new MailChimp_WooCommerce_ProductVariation();
             $variant->setId("deleted_{$id}");
             $variant->setTitle("deleted_{$id}");