+++ includes/processes/class-mailchimp-woocommerce-abstract-sync.php
@@ -16,7 +16,7 @@ abstract class MailChimp_WooCommerce_Abtstract_Sync extends WP_Job
     private $api;
 
     /**
-     * @var MailChimpApi
+     * @var MailChimp_WooCommerce_MailChimpApi
      */
     private $mc;
 
@@ -116,6 +116,9 @@ abstract class MailChimp_WooCommerce_Abtstract_Sync extends WP_Job
         $this->setData('sync.syncing', true);
         $this->setData('sync.started_at', time());
 
+        // flag the store as syncing
+        mailchimp_get_api()->flagStoreSync(mailchimp_get_store_id(), true);
+
         return $this;
     }
 
@@ -124,11 +127,14 @@ abstract class MailChimp_WooCommerce_Abtstract_Sync extends WP_Job
      */
     public function flagStopSync()
     {
+        mailchimp_log('sync.completed', "Finished Sync :: ".date('D, M j, Y g:i A'));
+
         // this is the last thing we're doing so it's complete as of now.
         $this->setData('sync.syncing', false);
         $this->setData('sync.completed_at', time());
 
-        mailchimp_log('sync.completed', "Finished Sync :: ".date('D, M j, Y g:i A'));
+        // flag the store as sync_finished
+        mailchimp_get_api()->flagStoreSync(mailchimp_get_store_id(), false);
 
         return $this;
     }
@@ -311,12 +317,12 @@ abstract class MailChimp_WooCommerce_Abtstract_Sync extends WP_Job
     }
 
     /**
-     * @return MailChimpApi
+     * @return MailChimp_WooCommerce_MailChimpApi
      */
     protected function mailchimp()
     {
         if (empty($this->mc)) {
-            $this->mc = new MailChimpApi($this->getOption('mailchimp_api_key'));
+            $this->mc = new MailChimp_WooCommerce_MailChimpApi($this->getOption('mailchimp_api_key'));
         }
         return $this->mc;
     }